+++ tests/QueryType/ManagedResources/RequestBuilder/SynonymsTest.php
@@ -197,6 +197,7 @@ class SynonymsTest extends TestCase
         $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
+        // there's a bug since Solr 8.7 with HEAD requests if a term is set (SOLR-15116)
         $this->assertSame(Request::METHOD_GET, $request->getMethod());
         $this->assertSame('schema/analysis/synonyms/dutch/mad', $request->getHandler());
         $this->assertNull($request->getRawData());
@@ -211,7 +212,7 @@ class SynonymsTest extends TestCase
         $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
-        $this->assertSame(Request::METHOD_GET, $request->getMethod());
+        $this->assertSame(Request::METHOD_HEAD, $request->getMethod());
         $this->assertSame('schema/analysis/synonyms/dutch', $request->getHandler());
         $this->assertNull($request->getRawData());
     }